requirements.txt
gradio
torch
torchvision
Pillow
numpy
huggingface_hub